// src/app/auth/types/authState.interface.ts
import { BackendErrorsInterface } from "src/app/shared/types/backendErrors.interface";
import { CurrentUserInterface } from "src/app/shared/types/currentUser.interface";

export interface AuthStateInterface {
    isSubmitting: boolean;
    currentUser: CurrentUserInterface | null;
    isLoading: boolean;
    isLoggedIn: boolean | null;
    validationErrors: BackendErrorsInterface | null;
}